+++ R/samplesize_scABEL.R
@@ -127,7 +127,7 @@ sampleN.scABEL1 <- function(alpha=0.05, targetpower=0.8, theta0, theta1,
     cat("            Sample size estimation\n")
     cat("   (simulation based on ANOVA evaluation)\n")
     cat("---------------------------------------------\n")
-    cat("Study design: ",desi,"\n")
+    cat("Study design:",desi,"\n")
     cat("log-transformed data (multiplicative model)\n")
     cat(nsims,"studies for each step simulated.\n\n")
     cat("alpha  = ", alpha,", target power = ", targetpower,"\n", sep="")